requirements.txt
aioredis==1.1.0
alembic==0.9.9
asgi-redis==1.4.3
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
bcolz==0.12.1
Bottleneck==1.2.1
certifi==2018.4.16
channels==2.1.2
channels-redis==2.2.1
chardet==3.0.4
click==6.7
constantly==15.1.0
contextlib2==0.5.5
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.28.3
daphne==2.2.0
decorator==4.3.0
Django==2.0.6
djangorestframework==3.8.2
empyrical==0.5.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
intervaltree==2.1.0
Jinja2==2.10
kiwisolver==1.0.1
Logbook==1.4.0
lru-dict==1.1.6
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
mpld3==0.3
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
networkx==1.11
numexpr==2.6.5
numpy==1.14.5
pandas==0.18.1
pandas-datareader==0.5.0
patsy==0.5.0
pyparsing==2.2.0
pypiwin32==223
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
pywin32==223
redis==2.10.6
requests==2.19.1
requests-file==1.4.3
requests-ftp==0.3.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sortedcontainers==2.0.4
SQLAlchemy==1.2.8
statsmodels==0.9.0
tables==3.4.4
toolz==0.9.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.23
websocket-client==0.48.0
zipline==1.2.0
zope.interface==4.5.0
